+++ src/utils/errors/index.js
@@ -1,0 +1,5 @@
+module.exports.BaseError = require('./BaseError');
+module.exports.NotFoundError = require('./NotFoundError');
+module.exports.AuthError = require('./AuthError');
+module.exports.ForbiddenError = require('./ForbiddenError');
+module.exports.BadRequestError = require('./BadRequestError');