+++ CHANGELOG.md
@@ -1,10 +1,10 @@
 # Changelog
 
 All notable changes to this project will be documented in this file. 
-See [standard-version](https://github.com/conventional-changelog/standard-version) for commit guidelines.
 
-## 1.0.0 (Pending)
+## 1.0.0 (Not Released Yet)
 
 ### Features
 
-- initial release
+- Authentication and authorization using passport
+- Roles and permissions